Refactored keyboard layout data code
